[FIX] business_requirement: fix comments on GitHub, modify business.py, add wizards folder and file mail_compose_message.py
