+++ business_requirement/wizards/__init__.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
 # © 2016 Elico Corp (https://www.elico-corp.com).
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import models, wizards
-from . import tests
+from . import mail_compose_message
